+++ ui/src/config/About.vue
@@ -18,6 +18,14 @@
       </div>
     </div>
     <div v-loading.fullscreen.lock="actionLoading" class="flex" style="margin-bottom: 8px;">
+      <popconfirm title="你确定要退出吗?" @confirm="logout">
+        <template #reference>
+          <el-button type="danger" bg text icon="Back">
+            退出
+          </el-button>
+        </template>
+      </popconfirm>
+      <div style="margin: 6px;"></div>
       <popconfirm title="你确定重启吗?" @confirm="stop(0)">
         <template #reference>
           <el-button bg icon="RefreshRight" text type="warning">重启</el-button>
@@ -90,6 +98,7 @@ import MarkdownItGitHubAlerts from 'markdown-it-github-alerts'
 import 'markdown-it-github-alerts/styles/github-colors-light.css'
 import 'markdown-it-github-alerts/styles/github-colors-dark-media.css'
 import 'markdown-it-github-alerts/styles/github-base.css'
+import {Back} from "@element-plus/icons-vue";
 
 let md = markdownit({
   html: true,
@@ -149,6 +158,11 @@ onMounted(() => {
         about.value = res.data
       })
 })
+
+let logout = () => {
+  localStorage.removeItem('authorization')
+  location.reload()
+}
 
 let openUrl = (url) => window.open(url)
